add sessionService, authController to constructor; rm password hashing and storing; add external auth service for this

+++ packages/backend-logic/src/models/User/implementations/UserService.ts
@@ -1,12 +1,12 @@
-import type { UserDT, UserDTN } from '@m-cafe-app/models';
+import type { AuthResponse, UserDT, UserDTN, UserDTU, UserUniqueProperties } from '@m-cafe-app/models';
 import type { IUserService, IUserRepo } from '../interfaces';
-import type { ISessionService } from '../../Session/interfaces';
+import type { ISessionService } from '../../Session';
+import type { IAuthControllerInternal } from '../../Auth';
 import type { AdministrateUserBody } from '@m-cafe-app/utils';
 import { User } from '@m-cafe-app/models';
-import { CredentialsError, PasswordLengthError, ProhibitedError, hasOwnProperty } from '@m-cafe-app/utils';
+import { ApplicationError, AuthServiceError, BannedError, CredentialsError, PasswordLengthError, ProhibitedError, UnknownError, hasOwnProperty } from '@m-cafe-app/utils';
 import { UserMapper } from '../infrastructure';
 import { maxPasswordLen, minPasswordLen, possibleUserRights } from '@m-cafe-app/shared-constants';
-import bcryptjs from 'bcryptjs';
 import config from '../../../config.js';
 import logger from '../../../utils/logger';
 
@@ -14,7 +14,8 @@ import logger from '../../../utils/logger';
 export class UserService implements IUserService {
   constructor(
     readonly repo: IUserRepo,
-    readonly sessionService: ISessionService
+    readonly sessionService: ISessionService,
+    readonly authController: IAuthControllerInternal
   ) {}
 
   async getAll(): Promise<UserDT[]> {
@@ -36,8 +37,6 @@ export class UserService implements IUserService {
   }
 
   async getById(id: number): Promise<UserDT> {
-    // Apply Auth check!
-
     const user = await this.repo.getById(id);
 
     const res: UserDT = UserMapper.domainToDT(user);
@@ -53,8 +52,7 @@ export class UserService implements IUserService {
     return res;
   }
 
-  // async create(userDTN: UserDTN, password: string): Promise<UserDT> {
-  async create(userDTN: UserDTN): Promise<UserDT> {
+  async create(userDTN: UserDTN, userAgent: string): Promise<{ user: UserDT, auth: AuthResponse}> {
     if (
       userDTN.password === undefined
       ||
@@ -62,39 +60,114 @@ export class UserService implements IUserService {
     )
       throw new PasswordLengthError(`Password must be longer than ${minPasswordLen} and shorter than ${maxPasswordLen} symbols`);
 
-    const saltRounds = 10;
-    const passwordHash = await bcryptjs.hash(userDTN.password, saltRounds);
+    const savedUser = await this.repo.create(userDTN);
 
-    const savedUser = await this.repo.create(userDTN, passwordHash);
+    const { user: authorizedUser, auth: userAuth } = await this.resolveAuthLookupHashConflict(savedUser, userDTN.password);
 
-    const res: UserDT = UserMapper.domainToDT(savedUser);
+    if (!authorizedUser.rights)
+      throw new ApplicationError('Failed to create user');
+
+    await this.sessionService.create(authorizedUser.id, userAuth.token, userAgent, authorizedUser.rights);
+
+    const res: { user: UserDT, auth: AuthResponse } = {
+      user: UserMapper.domainToDT(authorizedUser),
+      auth: userAuth
+    };
     
     return res;
   }
 
-  async update(userDT: UserDT, password: string, newPassword?: string): Promise<UserDT> {
+  async update(userDTU: UserDTU, userAgent: string): Promise<{ user: UserDT, auth: AuthResponse}> {
 
-    const { passwordHash } = await this.repo.getPasswordHashRights(userDT.id);
+    const foundUser = await this.repo.getById(userDTU.id);
 
-    const passwordCorrect = await bcryptjs.compare(password, passwordHash);
+    if (foundUser.phonenumber === config.SUPERADMIN_PHONENUMBER)
+      throw new ProhibitedError('Attempt to alter superadmin');
+    if (foundUser.rights === 'disabled')
+      throw new ProhibitedError('Attempt to alter disabled user');
+    if (!foundUser.lookupHash || !foundUser.rights)
+      throw new ApplicationError('User data corrupt: lookupHash or rights are missing');
 
-    if (!passwordCorrect) throw new CredentialsError('Password incorrect');
+    const checkCredentials = await this.authController.verifyCredentials({
+      lookupHash: foundUser.lookupHash,
+      password: userDTU.password
+    });
 
-    let newPasswordHash = passwordHash;
-
-    if (newPassword) {
-      if (!(minPasswordLen < newPassword.length && newPassword.length < maxPasswordLen))
-        throw new PasswordLengthError(`Password must be longer than ${minPasswordLen} and shorter than ${maxPasswordLen} symbols`);
-
-      const saltRounds = 10;
-      newPasswordHash = await bcryptjs.hash(newPassword, saltRounds);
+    if (!checkCredentials.success || checkCredentials.error !== '') {
+      console.log(checkCredentials);
+      // ACHTUNG CHECK actual error texts coming from AuthController
+      if (checkCredentials.error === 'User not found')
+        // HERE! Change error to createAuthRequest instead after actual error codes investigation
+        throw new ApplicationError('User not found on auth server. Please, contact the admins to resolve this problem');
+        // HERE! Change error to createAuthRequest instead after actual error codes investigation
+      else if (checkCredentials.error === 'Wrong password')
+        throw new CredentialsError('Password incorrect');
+      else 
+        throw new UnknownError(checkCredentials.error);
+        // ACHTUNG CHECK actual error texts coming from AuthController
     }
 
-    const updatedUser = await this.repo.update(UserMapper.dtToDomain(userDT), newPasswordHash);
+    if (userDTU.newPassword)
+      if (!(minPasswordLen < userDTU.newPassword.length && userDTU.newPassword.length < maxPasswordLen))
+        throw new PasswordLengthError(`Password must be longer than ${minPasswordLen} and shorter than ${maxPasswordLen} symbols`);
 
-    const res: UserDT = UserMapper.domainToDT(updatedUser);
+    const updatedUser = await this.repo.update(UserMapper.dtToDomain(userDTU));
+
+    if (!updatedUser.lookupHash || !updatedUser.rights)
+      throw new ApplicationError('User data corrupt: lookupHash or rights are missing');
+
+    // Does not need to resolve any lookupHash conflicts because they should not exist -
+    // user must be already authorized to even get to this point
+    const userAuth = await this.authController.update({
+      id: updatedUser.id,
+      lookupHash: updatedUser.lookupHash,
+      oldPassword: userDTU.password,
+      newPassword: userDTU.newPassword ? userDTU.newPassword : userDTU.password
+    });
+
+    await this.sessionService.update(updatedUser.id, userAuth.token, userAgent, updatedUser.rights);
+
+    const res: { user: UserDT, auth: AuthResponse } = {
+      user: UserMapper.domainToDT(updatedUser),
+      auth: userAuth
+    };
     
     return res;
+  }
+
+  async authenticate(password: string, credential: UserUniqueProperties): Promise<{ user: UserDT; auth: AuthResponse; }> {
+
+    if (credential.phonenumber === config.SUPERADMIN_PHONENUMBER) throw new ProhibitedError('Superadmin must login only with a username');
+
+    const whereClause = {} as UserUniqueProperties;
+    if (credential.username) whereClause.username = credential.username;
+    if (credential.phonenumber) whereClause.phonenumber = credential.phonenumber;
+    if (credential.email) whereClause.email = credential.email;
+
+    const user = await this.repo.getByUniqueProperties(whereClause);
+
+    if (!user.lookupHash)
+      throw new ApplicationError('User data corrupt: lookupHash is missing. Please, contact the admins to resolve this problem');
+    if (user.rights === 'disabled')
+      throw new BannedError('Your account have been banned. Contact admin to unblock account');
+    if (user.deletedAt)
+      throw new ProhibitedError('You have deleted your own account. To delete it permanently or restore it, contact admin');
+
+    const auth = await this.authController.grant({ id: user.id, lookupHash: user.lookupHash, password });
+
+    if (!auth.token || auth.error || auth.id !== user.id)
+      throw new AuthServiceError(auth.error);
+
+    const res: { user: UserDT; auth: AuthResponse; } = {
+      user: UserMapper.domainToDT(user),
+      auth
+    };
+
+    return res;
+  }
+
+  async logout(id: number, userAgent: string): Promise<void> {
+    await this.sessionService.remove({ where: { userId: id, userAgent } });
   }
 
   async administrate(id: number, body: AdministrateUserBody): Promise<UserDT> {
@@ -124,12 +197,7 @@ export class UserService implements IUserService {
       }
 
       if (userSubject.rights === 'disabled') {
-        // Implement Session as Auth microservice!
-        // await Session.destroy({
-        //   where: {
-        //     userId: userSubject.id,
-        //   }
-        // });
+        await this.sessionService.remove({ where: { userId: userSubject.id } });
       }
     }
 
@@ -142,10 +210,16 @@ export class UserService implements IUserService {
    * Mark user as deleted, add deletedAd timestamp.
    */
   async remove(id: number): Promise<UserDT> {
-    // Auth module implement!
-    // await Session.destroy({ where: { userId: req.user.id } });
+    await this.sessionService.remove({ where: { userId: id } });
 
-    const res: UserDT = UserMapper.domainToDT(await this.repo.remove(id));
+    const deletedUser = await this.repo.remove(id);
+
+    if (!deletedUser.lookupHash)
+      throw new ApplicationError('User data corrupt: lookupHash is missing');
+
+    await this.authController.remove({ lookupHash: deletedUser.lookupHash });
+
+    const res: UserDT = UserMapper.domainToDT(deletedUser);
 
     return res;
   }
@@ -182,30 +256,44 @@ export class UserService implements IUserService {
         // Some paranoid checks
         if (
           existingUser.rights !== 'admin'
-        ||
-        existingUser.phonenumber !== config.SUPERADMIN_PHONENUMBER
+          ||
+          existingUser.phonenumber !== config.SUPERADMIN_PHONENUMBER
         ) {
           await this.repo.delete(existingUser.id);
           await this.initSuperAdmin();
         }
       }
     } catch (e) {
-      // Do nothing;
+      logger.shout('Could not update superadmin rights', e);
+      process.exit(1);
     }
-  
+
+    const superAdminPassword = process.env.SUPERADMIN_PASSWORD;
+    if (!superAdminPassword) 
+      throw new ApplicationError('Super admin password is not defined');
+
+    const superAdminUsername = process.env.SUPERADMIN_USERNAME;
+    if (!superAdminUsername) 
+      throw new ApplicationError('Super admin username is not defined');
+
     // New superadmin creation
-    const saltRounds = 10;// REMOVE THIS AFTER AUTH MODULE IS FINISHED
-    const passwordHash = await bcryptjs.hash(process.env.SUPERADMIN_PASSWORD as string, saltRounds);// REMOVE THIS AFTER AUTH MODULE IS FINISHED
-  
     const user = {
-      username: process.env.SUPERADMIN_USERNAME as string,
+      username: superAdminUsername,
       phonenumber: config.SUPERADMIN_PHONENUMBER,
-      password: process.env.SUPERADMIN_PASSWORD as string, // REMOVE THIS AFTER AUTH MODULE IS FINISHED
-      passwordHash, // REMOVE THIS AFTER AUTH MODULE IS FINISHED
+      password: superAdminPassword,
       rights: 'admin'
     };
   
-    await this.repo.create(user, passwordHash, true, 'admin');
+    const savedSuperAdmin = await this.repo.create(user, true, 'admin');
+
+    if (!savedSuperAdmin.lookupHash || !savedSuperAdmin.rights)
+      throw new ApplicationError('Failed to create superadmin');
+
+    await this.authController.create({
+      id: savedSuperAdmin.id,
+      lookupHash: savedSuperAdmin.lookupHash,
+      password: superAdminPassword
+    });
   
     process.env.SUPERADMIN_USERNAME = '';
     process.env.SUPERADMIN_PASSWORD = '';
@@ -213,4 +301,25 @@ export class UserService implements IUserService {
     return logger.info('Super admin user successfully created!');
   }
 
+  private async resolveAuthLookupHashConflict(user: User, password: string, tries: number = 0): Promise<{ user: User; auth: AuthResponse; }> {
+
+    if (tries > 5) throw new ApplicationError('Too many tries. This can happen once in an enormous amount of times. Interrupted to avoid data loss. Please, try again manually.');
+
+    if (!user.lookupHash)
+      throw new ApplicationError('User data corrupt: lookupHash is missing');
+    
+    const auth = await this.authController.create({ id: user.id, lookupHash: user.lookupHash, password });
+
+    if (!auth.token || auth.error || auth.id !== user.id) {
+      logger.shout('Failed to create auth token', auth);  // ACHTUNG! REMOVE THIS WHEN ACTUAL ERROR TEXT IS KNOWN
+      if (auth.error === 'User already exists') {
+        const userWithNewLookupHash = await this.repo.updateLookupHash(user, user.lookupNoise);
+        return this.resolveAuthLookupHashConflict(userWithNewLookupHash, password);
+      } else {
+        throw new UnknownError(auth.error);
+      }
+    }
+
+    return { user, auth };
+  }
 }
